@@ -4,15 +4,15 @@
     <parent>
         <groupId>be.bosa.dt</groupId>
         <artifactId>cbetools-parent</artifactId>
-        <version>1.5.2</version>
+        <version>1.5.3</version>
     </parent>
     <groupId>be.bosa.dt</groupId>
     <artifactId>cbetools-common</artifactId>
-	<version>1.5.2</version>
+	<version>1.5.3</version>
     <packaging>jar</packaging>
 	<properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 	<licenses>
 		<license>
@@ -49,7 +49,7 @@
         <dependency>
             <groupId>com.opencsv</groupId>
             <artifactId>opencsv</artifactId>
-            <version>5.7.1</version>
+            <version>5.8</version>
         </dependency>
     </dependencies>
     <name>cbetools-common</name>
